BreaksDebug refac
