+++ Breaks6502/BreaksDebug/BreaksDebug.sln
@@ -13,6 +13,8 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "M6502Core", "..\M6502Core\S
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "BaseLogicLib", "..\..\Common\BaseLogicLib\Scripts\VS2022\BaseLogicLib.vcxproj", "{11AAD192-46EB-4D5D-B81F-BCEE11D6AF8E}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SharpTools", "..\..\Common\SharpTools\SharpTools.csproj", "{B63A28E1-5C32-4164-934A-F163BD7589A3}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -77,6 +79,18 @@ Global
 		{11AAD192-46EB-4D5D-B81F-BCEE11D6AF8E}.Release|x64.Build.0 = Release|x64
 		{11AAD192-46EB-4D5D-B81F-BCEE11D6AF8E}.Release|x86.ActiveCfg = Release|Win32
 		{11AAD192-46EB-4D5D-B81F-BCEE11D6AF8E}.Release|x86.Build.0 = Release|Win32
+		{B63A28E1-5C32-4164-934A-F163BD7589A3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B63A28E1-5C32-4164-934A-F163BD7589A3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B63A28E1-5C32-4164-934A-F163BD7589A3}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B63A28E1-5C32-4164-934A-F163BD7589A3}.Debug|x64.Build.0 = Debug|Any CPU
+		{B63A28E1-5C32-4164-934A-F163BD7589A3}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B63A28E1-5C32-4164-934A-F163BD7589A3}.Debug|x86.Build.0 = Debug|Any CPU
+		{B63A28E1-5C32-4164-934A-F163BD7589A3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B63A28E1-5C32-4164-934A-F163BD7589A3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B63A28E1-5C32-4164-934A-F163BD7589A3}.Release|x64.ActiveCfg = Release|Any CPU
+		{B63A28E1-5C32-4164-934A-F163BD7589A3}.Release|x64.Build.0 = Release|Any CPU
+		{B63A28E1-5C32-4164-934A-F163BD7589A3}.Release|x86.ActiveCfg = Release|Any CPU
+		{B63A28E1-5C32-4164-934A-F163BD7589A3}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
